etcd: adjust loadPageLimit based on server feedback
RangeRequests are made in terms of number of entries, but must not
exceed the maximum number of bytes for a message. We don't know
the size of the entries so we have to guess. Rather than hardcode a
guess that might go wrong in the future, adjust the guess down based on
the error from the server.