etcd: rework GetRange to fix two safety issues
The first was just bad code: GetRange was grabbing a recursive read
lock on the DB mutex, making deadlocks possible.

The second is removing a rare race in higher-level usage of
GetRange. In cdb AddWatch, we call GetRange then grab the write
lock to add a watcher. This leaves a small window in which an update
can slip that the new watcher doesn't see. The GetRange final func
closes that window.