etcd: handle 'created' and 'fragment' watch messages
Sometimes etcd sends a 'created' messsage with a revision number we
haven't yet got the updates for, so it's important not to increment the
db.rev with its value. Similarly, a fragment message should not trigger
a db.rev update.

Caught by the etcd unit tests as soon as the incorrect optimization
for bypassing the watch update was disabled.