etcd: add an incremental load step when loadAll is slow
With large-ish ranges (3m KVs), our etcd cluster seems to get into a
state where sometimes a load can take ~10 seconds, or sometimes 2
minutes. This latter runs into compaction events meaning that before we
can start the watch, the key we want to watch from is no longer
available.

This commit adds a pass to load the delta KVs if a load took longer than
a minute, to reduce the likelihood we run into a compaction event.

(Why on Earth we are compacting so often is a mystery still to be
investigated.)