control/cfgdb: add PanicOnWrite
PanicOnWrite makes it easier to write tests that check
that a particular operation does not do any db writes.
This is useful for ensuring the unsecured endpoints
don't accidentally turn into a DOS vector.

Right now, one stuck client out there in the world
is writing 100k entries per day to the DB.