etcd: optimize WatchKey startup
Send the initial value of the watched key under a read lock.
Then before taking the write lock, check if that was enough for the
watch and exit early if it wasn't.

This also optimizes the typical case of the current value not changing
between the two locks, so processing the initial value happens under
the read lock.